organized dispose()

+++ core/src/com/nucleargame/items/RawUranium.java
@@ -15,8 +15,8 @@ public class RawUranium extends Item {
     }
     public void checkForProximity(){
         //calculate distance
-        distX=Math.abs(rect.x- Player.rect.x);
-        distY=Math.abs(rect.y- Player.rect.y);
+        distX=Math.abs((int)rect.x- Player.rect.x);
+        distY=Math.abs((int)rect.y- Player.rect.y);
         dist=(int)Math.sqrt(distX+distY);
 
         if(dist<=20){
